wbpost/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wbpost"
version = "0.1.0"
description = "Admin CLI that ties WB parser, bridge and maxapi into one auto-posting pipeline"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.5",
]

[project.scripts]
wbpost = "wbpost.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["wbpost*"]

[tool.ruff]
line-length = 110
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
asyncio_mode = "auto"
